+++ xmake/rules/asm/xmake.lua
@@ -20,7 +20,7 @@
 
 -- define rule: asm.build
 rule("asm.build")
-    set_extensions(".s", ".asm")    
+    set_sourcekinds("as")    
     on_build_files("private.action.build.object")
 
 -- define rule: asm